+++ app/src/main/res/layout/activity_drawing.xml
@@ -74,12 +74,6 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <Button
-        android:id="@+id/button2"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Button" />
-
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="0dp"
